Added askForChoice method to UserFeedback classes
You can now ask the user to choose a value from a dict. The user is presented
a choice dialog with all the keys. When the user selects an alternative askForChoice
returns the corresponding value. You must indicate the key of the default value.
This is selected when the user presses Enter or when playground is run with --no-ask