Added config file support for playground.
The config file is expected to be located in ~/.wns/playground.config
You can use this to have persistent configuration for plugins. Playground uses the
INI file format. To access the config entries playground uses the Python
config file parser ConfigParser.SafeConfigParser

To get the instance use core.getConfig(). You can modify the contents
of the INI file but do not forget to call core.updateConfig() to store your
modifications to disk and reread the config.